WIP introduce `get_layout_template_name()` - the Arm Layout way for a model to specify the templates it wants to use. This allows Embeds to use different templates based on response type. This follows the idea that photos render differently than links and videos. The type of embedded content matters when choosing the template.
This feature is finished and tested however it remains "WIP" because it relies on an incomplete PR for Arm Layouts :: https://github.com/armstrong/armstrong.core.arm_layout/pull/8
Therefore:
- there may be tweaks
- we need to add the package to the test requirements

Note: ArmLayouts is not required. If the package is not available, this feature is irrelevant and the mixin code becomes a "pass".